+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [Unreleased] - XXXX-XX-XX
+
+### Added
+
+- New namespace `influxdb2` for handler InfluxDB requests.
+
+### Changed
+
+### Fixed
+
+- Bug creating logs when a trial network is validated.
+
+### Removed
+
 ## [v0.5.0] - 2025-03-31
 
 ### Added
@@ -289,6 +303,7 @@
 
 - Frontend implementation.
 
+[Unreleased]: https://github.com/6G-SANDBOX/TNLCM/compare/v0.5.0...HEAD
 [v0.5.0]: https://github.com/6G-SANDBOX/TNLCM/compare/v0.4.5...v0.5.0
 [v0.4.5]: https://github.com/6G-SANDBOX/TNLCM/compare/v0.4.4...v0.4.5
 [v0.4.4]: https://github.com/6G-SANDBOX/TNLCM/compare/v0.4.3...v0.4.4